Fake Lag Packets Holder Manager
Allows for consolidated system for fake lag packets. Currently the latency seems to be off and does not show the true values of actual latency simulated by the fake lag packet manager

+++ Source/Utility/Game/GameManager.cpp
@@ -120,7 +120,7 @@ namespace Skel
 
 	// Returns delta time with scale
 	// Will return 0 if paused
-	float GameManager::DeltaTime() const
+	double GameManager::DeltaTime() const
 	{
 		// Pause / Glitch
 		if (m_DeltaTime > 1) {
@@ -131,7 +131,7 @@ namespace Skel
 
 	// Returns delta time ignoring time scale, eg. pause
 	// Useful for debug methods
-	float GameManager::DeltaTimeUnscaled() const
+	double GameManager::DeltaTimeUnscaled() const
 	{
 		return m_DeltaTime;
 	}
